refactor: move the identity reveal animation out of IdentityManager
animateReveal (scramble + glitch pass) was presentation living on the
identity state manager — and the reason user-fake-id still imported the
manager directly. It now lives in _helpers/identityReveal.ts as a
standalone function with module-scoped per-element bookkeeping; the
manager sheds ~220 lines (including the long-dead commented-out
animation drafts) and is now imported only by identity.ts and its own
tests. user-fake-id drops the IdentityManager import entirely — every
component reaches identity through the identity.ts seam alone.
