+++ v2-blog/src/_helpers/identityManager.ts
@@ -80,24 +80,6 @@ export class IdentityManager {
     return IdentityManager.instance;
   }
 
-  // public init(element: HTMLElement): void {
-  //   // const element = document.getElementById(elementId);
-  //   if (!element) return;
-  //   let seed = this.getCachedSeed();
-
-  //   if (!seed) {
-  //     const soul = this.getBrowserSoul();
-  //     seed = this.hashSoul(soul);
-  //     this.cacheSeed(seed);
-  //   }
-
-  //   const identityName = this.generateIdentity(seed);
-
-  //   requestAnimationFrame(() => {
-  //     this.animateReveal(element, identityName);
-  //   });
-  // }
-
   /**
    * Returns the full generated identity for the current browser session.
    * A stable seed is cached in session storage and reused across calls.
@@ -194,225 +176,6 @@ export class IdentityManager {
 
     return `${pre.toLocaleLowerCase()}${suf.toLocaleLowerCase()}::${id}`;
   }
-
-  // animation
-  // public animateReveal(element: HTMLElement, finalText: string): void {
-  //   const chars = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789@#$%^&*";
-  //   const duration = 800; // ms
-  //   const frames = 50; // steps
-  //   let step = 0;
-
-  //   // We create a custom animation loop for the text scrambling effect
-  //   // This is more performant than setInterval because it respects the browser refresh rate
-  //   const scramble = () => {
-  //     step++;
-  //     const progress = step / frames * duration;
-
-  //     // Calculate how many characters should be "revealed" (fixed) based on progress
-  //     const revealCount = Math.floor(finalText.length * progress);
-
-  //     const scrambledPart = finalText
-  //       .split('')
-  //       .map((char, index) => {
-  //         if (index < revealCount) return char; // Revealed char
-  //         if (char === ' ') return ' '; // Preserve spaces
-  //         return chars[Math.floor(Math.random() * chars.length)]; // random char
-  //       })
-  //       .join('');
-
-  //     element.innerText = scrambledPart;
-
-  //     if (step < frames) {
-  //       requestAnimationFrame(scramble);
-  //     } else {
-  //       // Ensure final state is clean
-  //       element.innerText = finalText;
-  //       this.triggerGlitchAnimation(element);
-  //     }
-  //   };
-
-  //   scramble();
-  // }
-
-  private _revealRaf = new WeakMap<HTMLElement, number>();
-
-  /**
-   * Cancels any pending reveal frame for an element and snaps its content to the final text.
-   *
-   * @param element - The element being updated by the reveal animation.
-   * @param finalText - The final resolved identity text.
-   */
-  private _snapToFinalText(element: HTMLElement, finalText: string): void {
-    const prev = this._revealRaf.get(element);
-    if (prev != null) {
-      cancelAnimationFrame(prev);
-      this._revealRaf.delete(element);
-    }
-
-    element.textContent = finalText;
-  }
-
-  /**
-   * Cancels any active glitch animation for an element and clears its temporary inline styles.
-   *
-   * @param element - The element that may own a glitch animation.
-   */
-  private _cancelGlitchAnimation(element: HTMLElement): void {
-    const prev = this._glitchAnim.get(element);
-    if (prev) {
-      prev.cancel();
-      this._glitchAnim.delete(element);
-    }
-
-    element.style.textShadow = "";
-    element.style.transform = "";
-  }
-
-  /**
-   * Reveals the final identity text through a scramble animation and then triggers the glitch pass.
-   * If the element disconnects, the text is snapped to the final value and stale work is canceled.
-   *
-   * @param element - The element whose text content should animate.
-   * @param finalText - The final identity string to reveal.
-   */
-  public animateReveal(element: HTMLElement, finalText: string): void {
-    if (!element) return;
-
-    const reduceMotion =
-      typeof window !== "undefined" &&
-      window.matchMedia?.("(prefers-reduced-motion: reduce)")?.matches;
-
-    if (reduceMotion) {
-      this._snapToFinalText(element, finalText);
-      this.triggerGlitchAnimation(element);
-      return;
-    }
-
-    const charset = "abcdefghijklmnopqrstuvwxyz0123456789";
-    const durationMs = 800;
-
-    const prev = this._revealRaf.get(element);
-    if (prev != null) cancelAnimationFrame(prev);
-    this._cancelGlitchAnimation(element);
-
-    const finalChars = Array.from(finalText); // handles unicode better than split('')
-    const len = finalChars.length;
-
-    const isSpace = finalChars.map(c => c === " ");
-
-    let start = 0;
-
-    /**
-     * Advances the scramble animation by one frame until the final text is fully revealed.
-     *
-     * @param now - The high-resolution RAF timestamp for the current frame.
-     */
-    const tick = (now: number) => {
-      if (!element.isConnected) {
-        this._snapToFinalText(element, finalText);
-        this._cancelGlitchAnimation(element);
-        return;
-      }
-
-      if (!start) start = now;
-      const elapsed = now - start;
-
-      // norm progress 0..1
-      const t = Math.min(1, elapsed / durationMs);
-      const eased = 1 - Math.pow(1 - t, 3); // easeoutcubic
-
-      const revealCount = Math.floor(len * eased);
-
-      // creates one string per frame
-      let out = "";
-      for (let i = 0; i < len; i++) {
-        if (i < revealCount) out += finalChars[i];
-        else if (isSpace[i]) out += " ";
-        else out += charset[(Math.random() * charset.length) | 0];
-      }
-
-      element.textContent = out;
-
-      if (t < 1) {
-        const id = requestAnimationFrame(tick);
-        this._revealRaf.set(element, id);
-      } else {
-        this._snapToFinalText(element, finalText);
-        this.triggerGlitchAnimation(element);
-      }
-    };
-
-    const id = requestAnimationFrame(tick);
-    this._revealRaf.set(element, id);
-  }
-
-
-  // private triggerGlitchAnimation(element: HTMLElement): void {
-  //   // Tailwind specific colors: text-red-500 (#ef4444) to your specific red (#ff453a)
-  //   const keyframes: Keyframe[] = [
-  //     { textShadow: '2px 0 #ff453a, -2px 0 #00ffff', transform: 'translate(2px, 0)' },
-  //     { textShadow: '-2px 0 #ff453a, 2px 0 #00ffff', transform: 'translate(-2px, 0)' },
-  //     { textShadow: '0 0 0', transform: 'translate(0, 0)' }
-  //   ];
-
-  //   element.animate(keyframes, {
-  //     duration: 500,
-  //     iterations: 3,
-  //     easing: 'steps(2, end)', // robotic movement
-  //   });
-  // }
-  private _glitchAnim = new WeakMap<HTMLElement, Animation>();
-
-  /**
-   * Runs the short glitch animation used after the reveal settles.
-   * The helper exits early for reduced-motion users and disconnected elements.
-   *
-   * @param element - The element that should receive the glitch effect.
-   */
-  private triggerGlitchAnimation(element: HTMLElement): void {
-    if (!element) return;
-
-    const reduceMotion =
-      typeof window !== "undefined" &&
-      window.matchMedia?.("(prefers-reduced-motion: reduce)")?.matches;
-
-    if (reduceMotion) return;
-
-    if (!element.isConnected) {
-      this._cancelGlitchAnimation(element);
-      return;
-    }
-
-    this._cancelGlitchAnimation(element);
-
-    const keyframes: Keyframe[] = [
-      { textShadow: "2px 0 #ff453a, -2px 0 #00ffff", transform: "translate(2px, 0)" },
-      { textShadow: "-2px 0 #ff453a, 2px 0 #00ffff", transform: "translate(-2px, 0)" },
-      { textShadow: "1px 0 #ff453a, -1px 0 #00ffff", transform: "translate(1px, 0)" },
-      { textShadow: "-1px 0 #ff453a, 1px 0 #00ffff", transform: "translate(-1px, 0)" },
-      { textShadow: "none", transform: "translate(0, 0)" },
-    ];
-
-    const anim = element.animate(keyframes, {
-      duration: 220,
-      iterations: 2,
-      easing: "steps(2, end)",
-      fill: "none",
-    });
-
-    this._glitchAnim.set(element, anim);
-
-    /**
-     * Clears the glitch bookkeeping and inline styles once the animation stops.
-     */
-    const cleanup = () => {
-      this._cancelGlitchAnimation(element);
-    };
-
-    anim.addEventListener("finish", cleanup, { once: true });
-    anim.addEventListener("cancel", cleanup, { once: true });
-  }
-
 
   /**
    * Reads the cached seed for this session from session storage.
