+++ v2-blog/src/components/user-fake-id.ts
@@ -1,24 +1,15 @@
 import { css, html, LitElement, nothing, PropertyValues } from "lit";
 import { customElement, property, query, state } from "lit/decorators.js";
 
-// IdentityManager is imported ONLY for animateReveal (presentation concern);
-// every identity-state read/write goes through the identity.ts seam.
-import { IdentityManager } from "../_helpers/identityManager.ts";
 import { getIdentity, getStoredIdentity, nextRandomIdentity, setIdentity } from "../_helpers/identity.ts";
+import { animateIdentityReveal } from "../_helpers/identityReveal.ts";
 
 @customElement('user-fakeid')
 export class UserFakeID extends LitElement {
-  constructor() {
-    super();
-    this.identity = IdentityManager.getInstance();
-  }
-
   @query('#userId') private _userElID!: HTMLSpanElement;
 
   @property({ type: String, reflect: true })
   userID: string = "";
-  // @property({ type: typeof IdentityManager })
-  // identity = IdentityManager.getInstance();
 
   @state() private isRandom = false;
   @state() private isDesktop = true;
@@ -28,7 +19,6 @@ export class UserFakeID extends LitElement {
     this.isDesktop = e.matches;
   };
 
-  private identity = IdentityManager.getInstance();
   private hasUserName = getStoredIdentity();
 
   static styles = css`
@@ -102,7 +92,7 @@ export class UserFakeID extends LitElement {
     if (!this.allowReveal) return;
 
     if (this.hasUserName) {
-      this.identity.animateReveal(this._userElID, this.hasUserName);
+      animateIdentityReveal(this._userElID, this.hasUserName);
       return;
     }
 
@@ -118,7 +108,7 @@ export class UserFakeID extends LitElement {
       this.updateComplete.then(() => {
         if (!this.allowReveal) return;
         if (!this._userElID) return;
-        this.identity.animateReveal(this._userElID, this.userID!);
+        animateIdentityReveal(this._userElID, this.userID!);
       })
     }
   }
@@ -126,7 +116,7 @@ export class UserFakeID extends LitElement {
   private async _handleRandomID(): Promise<void> {
     this.isRandom = true;
     this.userID = nextRandomIdentity();
-    await this.identity.animateReveal(this._userElID, this.userID);
+    animateIdentityReveal(this._userElID, this.userID);
   }
 
   private setRandomID(): void {
