+++ v2-blog/tests/unit/helpers/identityManager.test.ts
@@ -78,68 +78,6 @@ describe("identityManager", () => {
     }
   });
 
-  it("snaps to the final text if reveal work continues after disconnect", () => {
-    const manager = IdentityManager.getInstance();
-    const element = document.createElement("span");
-    document.body.appendChild(element);
-
-    const rafQueue: FrameRequestCallback[] = [];
-    const originalRaf = globalThis.requestAnimationFrame;
-    const originalCancel = globalThis.cancelAnimationFrame;
-
-    globalThis.requestAnimationFrame = vi.fn((cb: FrameRequestCallback) => {
-      rafQueue.push(cb);
-      return rafQueue.length;
-    }) as typeof requestAnimationFrame;
-    globalThis.cancelAnimationFrame = vi.fn() as typeof cancelAnimationFrame;
-
-    try {
-      manager.animateReveal(element, "echo_shell::1111");
-      element.remove();
-
-      const firstFrame = rafQueue.shift();
-      firstFrame?.(16);
-
-      expect(element.textContent).toBe("echo_shell::1111");
-    } finally {
-      globalThis.requestAnimationFrame = originalRaf;
-      globalThis.cancelAnimationFrame = originalCancel;
-    }
-  });
-
-  it("does not start glitch animation for a disconnected element after reveal snaps final text", () => {
-    const manager = IdentityManager.getInstance();
-    const element = document.createElement("span");
-    document.body.appendChild(element);
-
-    const animate = vi.fn();
-    element.animate = animate as unknown as typeof element.animate;
-
-    const rafQueue: FrameRequestCallback[] = [];
-    const originalRaf = globalThis.requestAnimationFrame;
-    const originalCancel = globalThis.cancelAnimationFrame;
-
-    globalThis.requestAnimationFrame = vi.fn((cb: FrameRequestCallback) => {
-      rafQueue.push(cb);
-      return rafQueue.length;
-    }) as typeof requestAnimationFrame;
-    globalThis.cancelAnimationFrame = vi.fn() as typeof cancelAnimationFrame;
-
-    try {
-      manager.animateReveal(element, "echo_shell::1111");
-      element.remove();
-
-      const firstFrame = rafQueue.shift();
-      firstFrame?.(16);
-
-      expect(animate).not.toHaveBeenCalled();
-      expect(element.textContent).toBe("echo_shell::1111");
-    } finally {
-      globalThis.requestAnimationFrame = originalRaf;
-      globalThis.cancelAnimationFrame = originalCancel;
-    }
-  });
-
   it("returns a stable default identity once a seed is cached", () => {
     const manager = IdentityManager.getInstance();
